move pulser comparison tests to single file
